refactor: 회원 기능 Auth, Domain 역할 분리
