+++ docs/Feedback.md
@@ -5,8 +5,10 @@
 - [ ] `테스트`
     - [ ] 시나리오 작성 후 해당 시나리오를 따라 테스트 코드 작성
     - [ ] Controller Test 추가
-- [ ] `Auth`
-    - [ ] `login` 이외의 기능이 인증, 인가에 관련된 부분인지 고민
+- [x] `Auth`
+    - [x] `login` 이외의 기능이 인증, 인가에 관련된 부분인지 고민
+      - `login` : Auth
+      - `이 외 기능` : Customer
 - [ ] `Exception`
     - [ ] 서버에서 에러가 발생한 경우 Bad Request 가 맞는지?
     - [ ] `ExceptionHandler` 코드 중복 제거
