+++ src/main/java/woowacourse/shoppingcart/ui/CustomerController.java
@@ -1,4 +1,57 @@
 package woowacourse.shoppingcart.ui;
 
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import woowacourse.auth.dto.CustomerRequest;
+import woowacourse.auth.dto.CustomerResponse;
+import woowacourse.auth.support.AuthenticationPrincipal;
+import woowacourse.shoppingcart.application.CustomerService;
+
+@RestController
+@RequestMapping("/customers")
 public class CustomerController {
+
+    private final CustomerService customerService;
+
+    public CustomerController(CustomerService customerService) {
+        this.customerService = customerService;
+    }
+
+    @PostMapping
+    public ResponseEntity<CustomerResponse> registerCustomers(@RequestBody final CustomerRequest customerRequest) {
+        CustomerResponse customerResponse = customerService.register(customerRequest);
+        return ResponseEntity.created(URI.create("/customers/login")).body(customerResponse);
+    }
+
+    @GetMapping
+    public ResponseEntity<CustomerResponse> findCustomer(@AuthenticationPrincipal Long id) {
+        CustomerResponse customerResponse = customerService.findCustomerById(id);
+        return ResponseEntity.ok().body(customerResponse);
+    }
+
+    @PutMapping
+    public ResponseEntity<Void> updateCustomer(@RequestBody CustomerRequest customerRequest,
+                                               @AuthenticationPrincipal Long id) {
+        customerService.edit(id, customerRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping
+    public ResponseEntity<Void> deleteCustomer(@AuthenticationPrincipal Long id) {
+        customerService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/email")
+    public ResponseEntity<Boolean> checkEmail(String email) {
+        boolean isValidEmail = customerService.validateEmail(email);
+        return ResponseEntity.ok().body(isValidEmail);
+    }
 }
